requirements.txt
bleak
pillow
